Updated product cards for Aave v3
