+++ blockchain/token-metadata-list/aave-v3/aave-v3-ethereum.ts
@@ -1,0 +1,143 @@
+import { MainNetworkNames } from 'blockchain/networks'
+import { TokenConfig } from 'blockchain/tokensMetadata'
+import { staticFilesRuntimeUrl } from 'helpers/staticPaths'
+import { LendingProtocol } from 'lendingProtocols'
+
+export const aaveV3EthereumTokensMetadata: TokenConfig[] = [
+  {
+    symbol: 'wstETHethV3',
+    precision: 18,
+    digits: 5,
+    digitsInstant: 2,
+    name: 'WSTETH / ETH',
+    icon: 'aave_wsteth_usdc',
+    iconCircle: 'aave_wsteth_usdc',
+    iconColor: 'aave_wsteth_usdc',
+    color: '#E2F7F9',
+    background: 'linear-gradient(160.47deg, #E2F7F9 0.35%, #D3F3F5 99.18%), #000000',
+    bannerIcon: staticFilesRuntimeUrl('/static/img/tokens/Aave_v3_stETH.png'),
+    bannerGif: staticFilesRuntimeUrl('/static/img/tokens/AAVE_stETH_v3.gif'),
+    tags: [],
+    protocol: LendingProtocol.AaveV3,
+    chain: MainNetworkNames.ethereumMainnet,
+  },
+  {
+    symbol: 'rethethV3',
+    precision: 18,
+    digits: 5,
+    digitsInstant: 2,
+    name: 'RETH / ETH',
+    icon: 'aave_wsteth_usdc',
+    iconCircle: 'aave_wsteth_usdc',
+    iconColor: 'aave_wsteth_usdc',
+    color: '#E2F7F9',
+    background: 'linear-gradient(160.47deg, #E2F7F9 0.35%, #D3F3F5 99.18%), #000000',
+    bannerIcon: staticFilesRuntimeUrl('/static/img/tokens/AAVE_rETH_v3.png'),
+    bannerGif: staticFilesRuntimeUrl('/static/img/tokens/AAVE_rETH_v3.gif'),
+    tags: [],
+    protocol: LendingProtocol.AaveV3,
+    chain: MainNetworkNames.ethereumMainnet,
+  },
+  {
+    symbol: 'cbethethV3',
+    precision: 18,
+    digits: 5,
+    digitsInstant: 2,
+    name: 'CBETH / ETH',
+    icon: 'aave_wsteth_usdc',
+    iconCircle: 'aave_wsteth_usdc',
+    iconColor: 'aave_wsteth_usdc',
+    color: '#E2F7F9',
+    background: 'linear-gradient(160.47deg, #E2F7F9 0.35%, #D3F3F5 99.18%), #000000',
+    bannerIcon: staticFilesRuntimeUrl('/static/img/tokens/AAVE_cbETH_v3.png'),
+    bannerGif: staticFilesRuntimeUrl('/static/img/tokens/AAVE_cbETH_v3.gif'),
+    tags: [],
+    protocol: LendingProtocol.AaveV3,
+    chain: MainNetworkNames.ethereumMainnet,
+  },
+  {
+    symbol: 'ethusdcV3',
+    precision: 18,
+    digits: 5,
+    digitsInstant: 2,
+    name: 'ETH / USDC',
+    icon: 'aave_eth_usdc',
+    iconCircle: 'aave_eth_usdc',
+    iconColor: 'aave_eth_usdc',
+    color: '#E2F7F9',
+    background: 'linear-gradient(160.47deg, #F0F3FD 0.35%, #FCF0FD 99.18%), #FFFFFF',
+    bannerIcon: staticFilesRuntimeUrl('/static/img/tokens/AAVE_ETH_v3.png'),
+    bannerGif: staticFilesRuntimeUrl('/static/img/tokens/AAVE_ETH_v3.gif'),
+    tags: [],
+    protocol: LendingProtocol.AaveV3,
+    chain: MainNetworkNames.ethereumMainnet,
+  },
+  {
+    symbol: 'cbETHusdcV3',
+    precision: 18,
+    digits: 5,
+    digitsInstant: 2,
+    name: 'cbETH / USDC',
+    icon: 'aave_cbETH_usdc',
+    iconCircle: 'aave_cbETH_usdc',
+    iconColor: 'aave_cbETH_usdc',
+    color: '#E2F7F9',
+    background: 'linear-gradient(160.47deg, #E0E8F5 0.35%, #F0FBFD 99.18%), #FFFFFF',
+    bannerIcon: staticFilesRuntimeUrl('/static/img/tokens/AAVE_cbETH_v3.png'),
+    bannerGif: staticFilesRuntimeUrl('/static/img/tokens/AAVE_cbETH_v3.gif'),
+    tags: [],
+    protocol: LendingProtocol.AaveV3,
+    chain: MainNetworkNames.ethereumMainnet,
+  },
+  {
+    symbol: 'wBTCusdcV3',
+    precision: 18,
+    digits: 5,
+    digitsInstant: 2,
+    name: 'WBTC / USDC',
+    icon: 'aave_wbtc_usdc',
+    iconCircle: 'aave_wbtc_usdc',
+    iconColor: 'aave_wbtc_usdc',
+    color: '#E2F7F9',
+    background: 'linear-gradient(147.66deg, #FEF1E1 0%, #FDF2CA 88.25%)',
+    bannerIcon: staticFilesRuntimeUrl('/static/img/tokens/AAVE_WBTC_v3.png'),
+    bannerGif: staticFilesRuntimeUrl('/static/img/tokens/AAVE_WBTC_v3.gif'),
+    tags: [],
+    protocol: LendingProtocol.AaveV3,
+    chain: MainNetworkNames.ethereumMainnet,
+  },
+  {
+    symbol: 'wstethusdcV3',
+    precision: 18,
+    digits: 5,
+    digitsInstant: 2,
+    name: 'WSTETH / USDC',
+    icon: 'aave_wsteth_usdc',
+    iconCircle: 'aave_wsteth_usdc',
+    iconColor: 'aave_wsteth_usdc',
+    color: '#E2F7F9',
+    background: 'linear-gradient(147.66deg, #FEF1E1 0%, #FDF2CA 88.25%)',
+    bannerIcon: staticFilesRuntimeUrl('/static/img/tokens/AAVE_stETH_V3.png'),
+    bannerGif: staticFilesRuntimeUrl('/static/img/tokens/AAVE_stETH_V3.gif'),
+    tags: [],
+    protocol: LendingProtocol.AaveV3,
+    chain: MainNetworkNames.ethereumMainnet,
+  },
+  {
+    symbol: 'rethusdcV3',
+    precision: 18,
+    digits: 5,
+    digitsInstant: 2,
+    name: 'RETH / USDC',
+    icon: 'aave_wsteth_usdc',
+    iconCircle: 'aave_wsteth_usdc',
+    iconColor: 'aave_wsteth_usdc',
+    color: '#E2F7F9',
+    background: 'linear-gradient(147.66deg, #FEF1E1 0%, #FDF2CA 88.25%)',
+    bannerIcon: staticFilesRuntimeUrl('/static/img/tokens/AAVE_rETH_V3.png'),
+    bannerGif: staticFilesRuntimeUrl('/static/img/tokens/AAVE_rETH_V3.gif'),
+    tags: [],
+    protocol: LendingProtocol.AaveV3,
+    chain: MainNetworkNames.ethereumMainnet,
+  },
+]
