+++ features/aave/strategy-config/all-actions-available.ts
@@ -1,0 +1,8 @@
+import { ManagePositionAvailableActions } from 'features/aave/common'
+
+export const allActionsAvailable: ManagePositionAvailableActions[] = [
+  'adjust',
+  'manage-debt',
+  'manage-collateral',
+  'close',
+]
