+++ features/aave/strategy-config/index.ts
@@ -1,0 +1,1 @@
+export * from './strategy-config'
