+++ helpers/useFeatureToggle.ts
@@ -35,8 +35,11 @@ export type Feature =
   | 'AaveV3Arbitrum'
   | 'AaveV3MultiplycbETHusdc'
   | 'AaveV3MultiplywBTCusdc'
-  | 'AaveV3MultiplywstETHeth'
+  | 'AaveV3MultiplywstETHusdc'
   | 'AaveV3MultiplyETHusdc'
+  | 'AaveV3MultiplyrETHusdc'
+  | 'AaveV3EarnrETHeth'
+  | 'AaveV3EarncbETHeth'
   | 'AaveV2ProductCard'
   | 'OasisCreate'
   | 'SwapWidget'
@@ -76,8 +79,11 @@ const configuredFeatures: Record<Feature, boolean> = {
   AaveV2ProductCard: true,
   AaveV3MultiplycbETHusdc: false,
   AaveV3MultiplywBTCusdc: false,
-  AaveV3MultiplywstETHeth: false,
   AaveV3MultiplyETHusdc: false,
+  AaveV3EarncbETHeth: false,
+  AaveV3MultiplyrETHusdc: false,
+  AaveV3EarnrETHeth: false,
+  AaveV3MultiplywstETHusdc: false,
   OasisCreate: false,
   SwapWidget: true,
   ProxyReveal: false,
