+++ lendingProtocols/aaveCommon/AaveServices.ts
@@ -4,7 +4,10 @@ import { Observable } from 'rxjs'
 import { LiquidationCallEvent } from 'types/ethers-contracts/AaveV2LendingPool'
 
 import { ProtocolData } from './protocolData'
-import { ReserveConfigurationData } from './reserveConfigurationData'
+import {
+  AaveReserveConfigurationDataParams,
+  ReserveConfigurationData,
+} from './reserveConfigurationData'
 import { ReserveData } from './reserveData'
 import { UserAccountData, UserAccountDataArgs } from './userAccountData'
 import { UserConfigurationResults } from './userConfiguration'
@@ -16,7 +19,9 @@ export interface AaveServices {
     proxyAddress: string,
   ) => Observable<ProtocolData>
   protocol: LendingProtocol
-  aaveReserveConfigurationData$: (args: { token: string }) => Observable<ReserveConfigurationData>
+  aaveReserveConfigurationData$: (
+    args: AaveReserveConfigurationDataParams,
+  ) => Observable<ReserveConfigurationData>
   getAaveReserveData$: (args: { token: string }) => Observable<ReserveData>
   getAaveAssetsPrices$: (args: { tokens: string[] }) => Observable<BigNumber[]>
   aaveAvailableLiquidityInUSDC$: (args: { token: string }) => Observable<BigNumber>
