+++ public/locales/en/common.json
@@ -1709,23 +1709,7 @@
         "title": "stETH/ETH Aave",
         "description": "Earn on your ETH with increased staking rewards. Borrow ETH on Aave to get more exposure to stETH."
       },
-      "stETHethV3": {
-        "title": "stETH/ETH Aave",
-        "description": "Earn on your ETH with increased staking rewards. Borrow ETH on Aave to get more exposure to stETH."
-      },
-      "wstETHethV2": {
-        "title": "wstETH/ETH Aave",
-        "description": "Enter with ETH, and get up to 9.99x Multiple on stETH staking rewards."
-      },
-      "wstETHethV3": {
-        "title": "wstETH/ETH Aave",
-        "description": "Enter with ETH, and get up to 9.99x Multiple on stETH staking rewards."
-      },
       "stETHusdcV2": {
-        "title": "Highest stETH multiple option",
-        "description": "Maximise your stETH exposure with Aave and USDC."
-      },
-      "stETHusdcV3": {
         "title": "Highest stETH multiple option",
         "description": "Maximise your stETH exposure with Aave and USDC."
       },
@@ -1733,26 +1717,43 @@
         "title": "Highest ETH multiple option",
         "description": "Maximise your ETH exposure with Aave and USDC."
       },
+      "wBTCusdcV2": {
+        "title": "High wBTC multiple option",
+        "description": "The biggest possible Multiple to maximize your exposure to WBTC"
+      },
+      "wstETHethV3": {
+        "title": "wstETH/ETH Aave",
+        "description": "Enter with ETH, and get up to 9.99x Multiple on stETH staking rewards."
+      },
+      "rethethV3": {
+        "title": "rETH/ETH Aave",
+        "description": "The biggest possible Multiple to maximize your exposure to rETH"
+      },
+      "cbethethV3": {
+        "title": "cbETH/ETH Aave",
+        "description": "The biggest possible Multiple to maximize your exposure to cbETH"
+      },
       "ethusdcV3": {
         "title": "Highest ETH multiple option",
         "description": "Maximise your ETH exposure with Aave and USDC."
       },
-      "wBTCusdcV2": {
-        "title": "High wBTC multiple option",
-        "description": "The biggest possible Multiple to maximize your exposure to WBTC"
+      "cbETHusdcV3": {
+        "title": "High cbETH multiple option",
+        "description": "Maximise your cbETH exposure with Aave and USDC."
       },
       "wBTCusdcV3": {
         "title": "High wBTC multiple option",
         "description": "The biggest possible Multiple to maximize your exposure to WBTC"
       },
-      "cbETHusdcV2": {
-        "title": "High cbETH multiple option",
-        "description": "The biggest possible Multiple to maximize your exposure to cbETH"
+      "wstethusdcV3": {
+        "title": "High wstETH multiple option",
+        "description": "Maximise your wstETH exposure with Aave and USDC."
       },
-      "cbETHusdcV3": {
-        "title": "High cbETH multiple option",
-        "description": "The biggest possible Multiple to maximize your exposure to cbETH"
+      "rethusdcV3": {
+        "title": "High rETH multiple option",
+        "description": "Maximise your rETH exposure with Aave and USDC."
       },
+
       "borrow-against-ETH": {
         "title": "AAVE ETH",
         "description": "A position to borrow against your ETH in AAVE"
